backend/requirements.txt
deepagents
langchain-openai
langchain-anthropic
pymongo
motor
python-dotenv
pydantic
